# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Musholla Nurul Iman API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events:
    get:
      operationId: getEvents
      tags: [events]
      summary: Get all events
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEvent"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{id}:
    put:
      operationId: updateEvent
      tags: [events]
      summary: Update event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEvent"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
  /gallery:
    get:
      operationId: getGallery
      tags: [gallery]
      summary: Get gallery items
      responses:
        "200":
          description: List of gallery items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GalleryItem"
    post:
      operationId: createGalleryItem
      tags: [gallery]
      summary: Create gallery item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGalleryItem"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryItem"
  /gallery/{id}:
    put:
      operationId: updateGalleryItem
      tags: [gallery]
      summary: Update gallery item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGalleryItem"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryItem"
    delete:
      operationId: deleteGalleryItem
      tags: [gallery]
      summary: Delete gallery item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
  /cctv:
    get:
      operationId: getCctvCameras
      tags: [cctv]
      summary: Get all CCTV cameras
      responses:
        "200":
          description: List of cameras
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CctvCamera"
    post:
      operationId: createCctvCamera
      tags: [cctv]
      summary: Create CCTV camera
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCctvCamera"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CctvCamera"
  /cctv/{id}:
    put:
      operationId: updateCctvCamera
      tags: [cctv]
      summary: Update CCTV camera
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCctvCamera"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CctvCamera"
    delete:
      operationId: deleteCctvCamera
      tags: [cctv]
      summary: Delete CCTV camera
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [settings]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginRequest"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResponse"
components:
  schemas:
...
    DeleteResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
...
    Event:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        location:
          type: string
        startDate:
          type: string
        endDate:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - title
        - description
        - location
        - startDate
        - isActive
        - createdAt
    CreateEvent:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        location:
          type: string
        startDate:
          type: string
        endDate:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        isActive:
          type: boolean
      required:
        - title
        - description
        - location
        - startDate
        - isActive
    GalleryItem:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
        title:
          type: string
        imageUrl:
          type: string
        category:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - type
        - title
        - imageUrl
        - category
        - isActive
        - createdAt
    CreateGalleryItem:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        imageUrl:
          type: string
        category:
          type: string
        isActive:
          type: boolean
      required:
        - type
        - title
        - imageUrl
        - category
        - isActive
    CctvCamera:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        location:
          type: string
        streamUrl:
          type: string
        embedUrl:
          type: string
          nullable: true
        isActive:
          type: boolean
        order:
          type: integer
        description:
          type: string
          nullable: true
      required:
        - id
        - name
        - location
        - streamUrl
        - isActive
        - order
    CreateCctvCamera:
      type: object
      properties:
        name:
          type: string
        location:
          type: string
        streamUrl:
          type: string
        embedUrl:
          type: string
          nullable: true
        isActive:
          type: boolean
        order:
          type: integer
        description:
          type: string
          nullable: true
      required:
        - name
        - location
        - streamUrl
        - isActive
        - order
...
    AdminLoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    AdminLoginResponse:
      type: object
      properties:
        success:
          type: boolean
        token:
          type: string
          nullable: true
        message:
          type: string
      required:
        - success
        - message