# lib/api-zod/src/generated/types/cctvCamera.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Musholla Nurul Iman API
 * OpenAPI spec version: 0.1.0
 */

export interface CctvCamera {
  id: number;
  name: string;
  location: string;
  streamUrl: string;
  embedUrl?: string | null;
  isActive: boolean;
  order: number;
  description?: string | null;
}