# lib/api-zod/src/generated/types/createGalleryItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Musholla Nurul Iman API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateGalleryItem {
  type: string;
  title: string;
  imageUrl: string;
  category: string;
  isActive: boolean;
}